+++ main.js
@@ -1,11 +1,11 @@
 // Automatic Table Of Contents 1.6.1 (https://github.com/johansatge)
-var c=(t,e)=>()=>(e||t((e={exports:{}}).exports,e),e.exports);var d=c((J,b)=>{var p=class{},g={},k=class{},v=t=>t;if(I()){let t=require("obsidian");p=t.Plugin,g=t.MarkdownRenderer,k=t.MarkdownRenderChild,v=t.htmlToMarkdown}b.exports={Plugin:p,MarkdownRenderer:g,MarkdownRenderChild:k,htmlToMarkdown:v};function I(){return typeof process!="object"?!0:!process.env||!process.env.JEST_WORKER_ID}});var M=c((Z,w)=>{var i={title:{type:"string",default:"",comment:""},style:{type:"value",default:"nestedList",values:["nestedList","nestedOrderedList","inlineFirstLevel"],comment:"TOC style (nestedList|nestedOrderedList|inlineFirstLevel)"},minLevel:{type:"number",default:0,comment:"Include headings from the specified level"},maxLevel:{type:"number",default:0,comment:"Include headings up to the specified level"},includeLinks:{type:"boolean",default:!0,comment:"Make headings clickable"},hideWhenEmpty:{type:"boolean",default:!1,comment:"Hide TOC if no headings are found"},debugInConsole:{type:"boolean",default:!1,comment:"Print debug info in Obsidian console"}};w.exports={getOptionsDocs:x,parseOptionsFromSourceText:T};function x(){let t=[];for(let e of Object.keys(i)){let n=i[e],r=n.comment.length>0?` # ${n.comment}`:"";t.push(`${e}: ${n.default}${r}`)}return t.join(`
-`)}function T(t=""){let e={};for(let n of Object.keys(i))e[n]=i[n].default;for(let n of t.split(`
-`)){let r=F(n);r!==null&&(e[r.name]=r.value)}return e}function F(t){let e=t.match(/([a-zA-Z0-9._ ]+):(.*)/);if(t.startsWith("#")||!e)return null;let n=e[1].trim(),r=i[n],o=e[2].trim();(!r||r.type!=="string")&&(o=o.replace(/#[^#]*$/,"").trim());let l=new Error(`Invalid value for \`${n}\``);if(r&&r.type==="number"){let s=Number.parseInt(o);if(s<0)throw l;return{name:n,value:s}}if(r&&r.type==="boolean"){if(!["true","false"].includes(o))throw l;return{name:n,value:o==="true"}}if(r&&r.type==="value"){if(!r.values.includes(o))throw l;return{name:n,value:o}}return r&&r.type==="string"?{name:n,value:o}:null}});var $=c((G,C)=>{var{htmlToMarkdown:H}=d();C.exports={getMarkdownFromHeadings:P};function P(t,e){let n={nestedList:A,nestedOrderedList:R,inlineFirstLevel:E},r="";if(e.title&&e.title.length>0){let l=e.style==="inlineFirstLevel"?" ":`
-`;r+=`${e.title}${l}`}let o=n[e.style](t,e);return o===null?e.hideWhenEmpty?"":`${r}_Table of contents: no headings found_`:r+o}function A(t,e){return y(t,!1,e)}function R(t,e){return y(t,!0,e)}function y(t,e,n){let r=e?"1.":"-",o=[],l=n.minLevel>0?n.minLevel:Math.min(...t.map(s=>s.level));for(let s of t)s.level<l||n.maxLevel>0&&s.level>n.maxLevel||s.heading.length!==0&&o.push(`${"	".repeat(s.level-l)}${r} ${L(s,n)}`);return o.length>0?o.join(`
-`):null}function E(t,e){let n=e.minLevel>0?e.minLevel:Math.min(...t.map(o=>o.level)),r=t.filter(o=>o.level===n).filter(o=>o.heading.length>0).map(o=>L(o,e));return r.length>0?r.join(" | "):null}function L(t,e){let n=s=>s.replaceAll("*","").replaceAll("_","").replaceAll("`","").replaceAll("==","").replaceAll("~~","").replace(/\[([^\]]+)\]\([^)]+\)/g,"$1"),r=s=>n(H(s)),o=(s,a)=>s.replace(/\[\[([^\]]+)\|([^\]]+)\]\]/g,a?"$1 $2":"$2").replace(/\[\[([^\]]+)\]\]/g,"$1").replaceAll("[[","").replaceAll("| ",a?"":"- ").replaceAll("|",a?" ":"-"),l=s=>s.replaceAll("#","");if(e.includeLinks){let s=t.heading;s=n(s),s=r(s),s=o(s,!1);let a=t.heading;return a=o(a,!0),a=l(a),`[[#${a}|${s}]]`}return t.heading}});var{Plugin:_,MarkdownRenderer:O,MarkdownRenderChild:j}=d(),{getOptionsDocs:S,parseOptionsFromSourceText:W}=M(),{getMarkdownFromHeadings:q}=$(),f="table-of-contents",N="toc",m=class extends _{async onload(){let e=(n,r,o)=>{o.addChild(new h(this.app,r,o.sourcePath,n))};this.registerMarkdownCodeBlockProcessor(f,e),this.registerMarkdownCodeBlockProcessor(N,e),this.addCommand({id:"insert-automatic-table-of-contents",name:"Insert table of contents",editorCallback:D}),this.addCommand({id:"insert-automatic-table-of-contents-docs",name:"Insert table of contents (with available options)",editorCallback:B})}};function D(t){let e=`\`\`\`${f}
-\`\`\``;t.replaceRange(e,t.getCursor())}function B(t){let e=[`\`\`\`${f}
-${S()}
+var l=(t,e)=>()=>(e||t((e={exports:{}}).exports,e),e.exports);var c=l((U,v)=>{var f=class{},p={},g=class{},k=t=>t;if(T()){let t=require("obsidian");f=t.Plugin,p=t.MarkdownRenderer,g=t.MarkdownRenderChild,k=t.htmlToMarkdown}v.exports={Plugin:f,MarkdownRenderer:p,MarkdownRenderChild:g,htmlToMarkdown:k};function T(){return typeof process!="object"?!0:!process.env||!process.env.JEST_WORKER_ID}});var b=l((ee,w)=>{var a={title:{type:"string",default:"",comment:""},style:{type:"value",default:"nestedList",values:["nestedList","nestedOrderedList","inlineFirstLevel"],comment:"TOC style (nestedList|nestedOrderedList|inlineFirstLevel)"},minLevel:{type:"number",default:0,comment:"Include headings from the specified level"},maxLevel:{type:"number",default:0,comment:"Include headings up to the specified level"},includeLinks:{type:"boolean",default:!0,comment:"Make headings clickable"},hideWhenEmpty:{type:"boolean",default:!1,comment:"Hide TOC if no headings are found"},debugInConsole:{type:"boolean",default:!1,comment:"Print debug info in Obsidian console"}};w.exports={getOptionsDocs:H,parseOptionsFromSourceText:P};function H(){let t=[];for(let e of Object.keys(a)){let n=a[e],r=n.comment.length>0?` # ${n.comment}`:"";t.push(`${e}: ${n.default}${r}`)}return t.join(`
+`)}function P(t=""){let e={};for(let n of Object.keys(a))e[n]=a[n].default;for(let n of t.split(`
+`)){let r=A(n);r!==null&&(e[r.name]=r.value)}return e}function A(t){let e=t.match(/([a-zA-Z0-9._ ]+):(.*)/);if(t.startsWith("#")||!e)return null;let n=e[1].trim(),r=a[n],o=e[2].trim();(!r||r.type!=="string")&&(o=o.replace(/#[^#]*$/,"").trim());let i=new Error(`Invalid value for \`${n}\``);if(r&&r.type==="number"){let s=Number.parseInt(o);if(s<0)throw i;return{name:n,value:s}}if(r&&r.type==="boolean"){if(!["true","false"].includes(o))throw i;return{name:n,value:o==="true"}}if(r&&r.type==="value"){if(!r.values.includes(o))throw i;return{name:n,value:o}}if(r&&r.type==="string")return{name:n,value:o};if(r&&r.type==="regexp")try{let s=new RegExp(o,"gi");return{name:n,value:s}}catch{throw i}return null}});var C=l((te,L)=>{var{htmlToMarkdown:R}=c();L.exports={getFormattedMarkdownHeading:E};function E(t,e){if(e.includeLinks){let n=t;n=y(n),n=_(n),n=M(n,!1);let r=t;return r=M(r,!0),r=j(r),`[[#${r}|${n}]]`}return t}function y(t){return t.replaceAll("*","").replaceAll("_","").replaceAll("`","").replaceAll("==","").replaceAll("~~","").replace(/\[([^\]]+)\]\([^)]+\)/g,"$1")}function _(t){return y(R(t))}function M(t,e){return t.replace(/\[\[([^\]]+)\|([^\]]+)\]\]/g,e?"$1 $2":"$2").replace(/\[\[([^\]]+)\]\]/g,"$1").replaceAll("[[","").replaceAll("| ",e?"":"- ").replaceAll("|",e?" ":"-")}function j(t){return t.replaceAll("#","")}});var I=l((ne,O)=>{var{getFormattedMarkdownHeading:$}=C();O.exports={getMarkdownFromHeadings:S};function S(t,e){let n={nestedList:q,nestedOrderedList:W,inlineFirstLevel:N},r="";if(e.title&&e.title.length>0){let i=e.style==="inlineFirstLevel"?" ":`
+`;r+=`${e.title}${i}`}let o=n[e.style](t,e);return o===null?e.hideWhenEmpty?"":`${r}_Table of contents: no headings found_`:r+o}function q(t,e){return x(t,!1,e)}function W(t,e){return x(t,!0,e)}function x(t,e,n){let r=e?"1.":"-",o=[],i=n.minLevel>0?n.minLevel:Math.min(...t.map(s=>s.level));for(let s of t)s.level<i||n.maxLevel>0&&s.level>n.maxLevel||s.heading.length!==0&&o.push(`${"	".repeat(s.level-i)}${r} ${$(s.heading,n)}`);return o.length>0?o.join(`
+`):null}function N(t,e){let n=e.minLevel>0?e.minLevel:Math.min(...t.map(o=>o.level)),r=t.filter(o=>o.level===n).filter(o=>o.heading.length>0).map(o=>$(o.heading,e));return r.length>0?r.join(" | "):null}});var{Plugin:D,MarkdownRenderer:F,MarkdownRenderChild:B}=c(),{getOptionsDocs:z,parseOptionsFromSourceText:J}=b(),{getMarkdownFromHeadings:K}=I(),h="table-of-contents",V="toc",u=class extends D{async onload(){let e=(n,r,o)=>{o.addChild(new m(this.app,r,o.sourcePath,n))};this.registerMarkdownCodeBlockProcessor(h,e),this.registerMarkdownCodeBlockProcessor(V,e),this.addCommand({id:"insert-automatic-table-of-contents",name:"Insert table of contents",editorCallback:Z}),this.addCommand({id:"insert-automatic-table-of-contents-docs",name:"Insert table of contents (with available options)",editorCallback:G})}};function Z(t){let e=`\`\`\`${h}
+\`\`\``;t.replaceRange(e,t.getCursor())}function G(t){let e=[`\`\`\`${h}
+${z()}
 \`\`\``];t.replaceRange(e.join(`
-`),t.getCursor())}var h=class extends j{constructor(e,n,r,o){super(n),this.app=e,this.element=n,this.sourcePath=r,this.sourceText=o}onload(){this.render(),this.registerEvent(this.app.metadataCache.on("changed",this.onMetadataChange.bind(this)))}onMetadataChange(){this.render()}render(){try{let e=W(this.sourceText);e.debugInConsole&&u("Options",e);let n=this.app.metadataCache.getCache(this.sourcePath),r=n?.headings?n.headings:[];e.debugInConsole&&u("Headings",r);let o=q(r,e);e.debugInConsole&&u("Markdown",o),this.element.empty(),O.renderMarkdown(o,this.element,this.sourcePath,this)}catch(e){let n=`_\u{1F4A5} Could not render table of contents (${e.message})_`;O.renderMarkdown(n,this.element,this.sourcePath,this)}}};function u(t,e){console.log(`%cAutomatic Table Of Contents %c${t}:
-`,"color: orange; font-weight: bold","font-weight: bold",e)}module.exports=m;
+`),t.getCursor())}var m=class extends B{constructor(e,n,r,o){super(n),this.app=e,this.element=n,this.sourcePath=r,this.sourceText=o}onload(){this.render(),this.registerEvent(this.app.metadataCache.on("changed",this.onMetadataChange.bind(this)))}onMetadataChange(){this.render()}render(){try{let e=J(this.sourceText);e.debugInConsole&&d("Options",e);let n=this.app.metadataCache.getCache(this.sourcePath),r=n?.headings?n.headings:[];e.debugInConsole&&d("Headings",r);let o=K(r,e);e.debugInConsole&&d("Markdown",o),this.element.empty(),F.renderMarkdown(o,this.element,this.sourcePath,this)}catch(e){let n=`_\u{1F4A5} Could not render table of contents (${e.message})_`;F.renderMarkdown(n,this.element,this.sourcePath,this)}}};function d(t,e){console.log(`%cAutomatic Table Of Contents %c${t}:
+`,"color: orange; font-weight: bold","font-weight: bold",e)}module.exports=u;
